added new pages
