+++ app/(routes)/layout.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react'
+
+const layout = ({ children }: { children: ReactNode }) => {
+  return (
+    <div>{children}</div>
+  )
+}
+
+export default layout
